Added an initial method for movement that works in some games
This method relies on moving the cursor by an extremely negative number to guarantee that it is at the top left corner when it starts to be moved around. Unfortunately, the touchpad must be disabled before and while this program is being run for it to work accurately, and that means buttons do not work. I could try to write a version where it is moved extremely negatively during each movement (loop) but I am not sure how feasible it is.